Remove lazy evaluation of properties
Initially the code took the approach that if a property required another cmd
line invocation, that we would defer until the client actually requested it.
Unfortunately, this doesn't work when we are changing objects and we need to do
a generic object difference between old and new as we were then querying the
lvm objects directly, so they would be the same.  For the generic object
difference code to produce accurate signals we need the existing object state
and the previous object state, so the code now incurs the hit of doing the
additional properties value look ups when the object is initialized.

Signed-off-by: Tony Asleson <tasleson@redhat.com>