Fix Lv.devices when more than 1 PE exists
When a LV PE segments are located on  more than 1 PV we will get
multiple lines for each LV if we specify columns which have
multiple values pertaining to PE info.  We could either modify the
parsing logic to handle multiple lines which refer to the same
thing or remove them and then query that information separately.

At the moment I am taking the easier but slower approach of doing
a separate query to retrieve the needed information.

Signed-off-by: Tony Asleson <tasleson@redhat.com>