Create different LV create methods for different types
LV creation is very flexible with a plethora of different creation
options.  Instead of having one method with tons of different
options we will create a methods for each type.  There are
pros/cons to this approach, but I don't see how we can make this
any simpler and maintain the creation options available.

Renamed: LvCreate -> LvCreateLinear
Added: LvCreateStriped

Signed-off-by: Tony Asleson <tasleson@redhat.com>