Use the lvm shell as the backend
Instead of running separate subprocess for every LVM command being perfomed it
is possible to run lvm shell, pass the commands to it and grab its output. This
way lvm doesn't provide the return/exit code, but its not checked for any other
value than 0 anywhere anyway so it can just be determined from the STDERR output
(if there was any or not).