Bug fix: lvs property on vg interface incorrect
If a lv is a thin pool we have a different interface for them.  The code
was generating the path to a typical lv, not a thinpool.  Pass the
attributes as well as the lv name back from lvs_in_vg so that we can create
the correct object path based on the lv type.

Signed-off-by: Tony Asleson <tasleson@redhat.com>