Make VG & LV object paths opaque
The paths are now simply an integer for each of the objects, eg.

/com/redhat/lvm1/pv/[0..n]
/com/redhat/lvm1/vg/[0..n]
/com/redhat/lvm1/lv/[0..n]
/com/redhat/lvm1/thinpool/[0..n]

To facilitate this we need a way to take a LVM name eg.

/dev/sda
some_vg
some_vg/lv1
some_other_vg/lv1

and make it into a consistent dbus object path.  Previously given the name
information we could easily determine the object path.  Now with the opaque
object paths we always need to look up a path given one of LV names.  As names
and uuids can be changed by the user it's difficult to keep a look up table in
a consistent state, especially through re-names.  Thus look ups are accomplished
by first checking by name and if that fails, by uuid.  This should work as
long as the two don't change at the same time.  The uuid would ideally be the
primary look up mechanism, but we need a way to allow users to do look ups given
simple names too.  Additionally at this time it's not possible to get the uuid
at the same time as you get the name for LVM assets that have values that point
to other LVM assets, eg. pool_lv in a lvs output.

On top of all this, the problem is complicated because we have circular
dependencies, each of the objects has references to other objects, thus when
we are constructing an object, the other objects might not exist.  Thus the
look up mechanism can also create 'future' place holder look ups.  This way we
can construct objects with the correct paths when they are created.  Previously,
I was using names and then generating the paths based on the name, but as names
can change you cannot store anything by name which is referring to some other
dbus object.

Even though it was a pain handling renames when the object path included them
the code ultimately was easier.  However, doing this we reduce the amount of
signals that are generated when a user does a simple rename.

Signed-off-by: Tony Asleson <tasleson@redhat.com>