Introduce method time-out handling on Manager methods
This commit introduces the ability for the client to control the length of time
a method is allowed to run before it returns a job.  The 'tmo' option works
as follows:

1. -1 = Client will wait until the operation is done.  Client will need to
increase their timeout to ensure that the dbus layer doesn't timeout before
the response is delivered.
2. 0 = Client is unwilling to wait, immediately return a job object path.
3. 1..N = Client is willing to wait up to the specified number of seconds
before a job object path is returned.  Make sure this timeout is less than
the dbus transport timeout configured by client.

Implementation details:
A queue and thread were added and a request object.  When a method is invoked
a request object is created which represents the request.  At the time this
request is created we will start a timer if the user selected option 3 above.
If the job completes before the timer expires the client gets the result.
If the timer expires before the request is completed we construct and return
a job path.

Each method that returned a path will now return (oo).  First object path is
the newly created object or '/' if not done.  Second object path is the job
object if a job was created.  They are mutually exclusive, when one is set
the other is '/'.  For other methods that don't return anything, they will
return '/' when complete signifying they are done, else the job path.

Signed-off-by: Tony Asleson <tasleson@redhat.com>