ExternalEvent: Handle update
After spending a lot of time trying to be selective in fetching only
the things I think have changed, I've come to the conclusion that
it's mostly a waste of time.  If you need to fetch 1 PV/VG/LV you
might as well fetch all as LVM time dominates service time.  However,
we do need to correct the load function so that if we are re-freshing
an existing object that we don't go back and query lvm as we already
have its state.  This will be done when I re-factor the load functions
to remove redundant/copied code.

Signed-off-by: Tony Asleson <tasleson@redhat.com>