Re-work dbus object re-fresh
Instead of forcing the removal and re-add of the dbus object we
are separating them into two.  This way we can replace the entire
state of a dbus object in one step and leave the dbus object
registered with the dbus library.

This change is fairly invasive, I'm guessing quite a bit is broken.
Need unit tests.  It also requires lvm version 2.02.132 for
operation, which is very new.  This is because we really need the
origin_lv uuid and the pool_lv uuid to correctly located existing
objects in memory when the uuid or name change in lvm.

Signed-off-by: Tony Asleson <tasleson@redhat.com>