lvmdbustest.py: Add test for timer expired path
This is a little tricky to test as we need to specify a timer value that is
shorter than the actual time it takes so that we test the timer expired
path.  As the timer is specified in seconds and usually lvm gets the job
done in a short amount of time this test may not pass consistently.  We
are trying to create a Vg and typically this completes in ~1.5 seconds on
the development VM.  This will likely not work on real hardware.  We try
up to 20 times to get the timer to expire before we fail the test.

Other possibilities include:

- Loop creating LVs until the time to create one exceeds 2 seconds and
then try to create one with a 1 second timeout
- Change the timeout to ms, so that we could specify a very short
timeout and make it easier to test, this would make testing quicker
and more consistent at the expense of buring more CPU cycles with
high resolution timer (according to docs)

Signed-off-by: Tony Asleson <tasleson@redhat.com>