Remove redundant refreshes
With the addition of the udev handling the unit test run-time went
from ~88 to ~255 seconds.  This was because we are getting a lot of
udev events which end up causing a lot of state refreshes.  Currently
we enqueue a refresh for every udev we determine we are interested in.
Thus we could have a number of re-freshes queued up which doesn't
improve things as they can all be handled by one at the end.

There are a number of approaches to handling this. Having a
timer that repeatedly gets extended by each re-fresh request until
they stop coming in causing the timer to expire causing an actual
refresh seems ideal.  However I've put in a simpler method of comparing
time stamps of when a request comes in to when the last one was done.  If
the request is older than the last refresh time we skip the refresh request.

I will work on the timer based approach later.

With this change we drop total test time to 121 seconds.

Signed-off-by: Tony Asleson <tasleson@redhat.com>