Remove RAID thinpool creation
The lvm command line does not support this, so the dbus service
was implementing this functionaliy by executing 3 separate lvm
command to accomplish the single API call.  The lvm team did not
like this approach and requested that I remove it and allow lvm
to do this internally instead.

RFE BZ: https://bugzilla.redhat.com/show_bug.cgi?id=1279636

Signed-off-by: Tony Asleson <tasleson@redhat.com>