AutomatedProperties.interface refactor
Previously we had the thinpool as an additional interface on an
lv.  To make this work we used some private data of the
dbus.service.Object, _dbus_interface_table which contained all of
the interfaces for the object.  This was removed from the latest
package for fedora causing the service to fail.

ref:
http://pkgs.fedoraproject.org/cgit/dbus-python.git/commit/\
?id=718250f05adee0152332e0528396a96001eab8a4

As we are not implementing multiple interfaces for any of the objects
we are presenting his commit simply removes the interface for which
the object was registered.  Later if we find that we need to support
multiple interfaces for the same object we will modify our code
to accomodate them instead of relying on the internal details of the
dbus.service.Object.

As always it would be better if the dbus service did not use private
data or methods as they are private and subject to change.

Signed-off-by: Tony Asleson <tasleson@redhat.com>