Re-work pv move handling
Instead of using "-b" to put the move into the background we
are leaving it and using a thread to grab stdout and parse it
while it's running.  This seems like an easier way and at the
moment we don't have the information we need to make using "-b"
work.

Signed-off-by: Tony Asleson <tasleson@redhat.com>