Make test_job_handler_timer pass
We were failing from a combination of a couple of things:

1. The operation really wasn't taking long enough
2. We had lock contention on the object manager lock

We solved 1 by creating a test VG and adding LVs to is so that the VG
remove takes longer than 1 second.  The second issue was resolved by
removing the object manager lock we were holding while scanning the
entire system for changes.  This doens't solve anything as external
users can change the state without our knowledge too, so removing it
as it really isn't buying us anything with that.

Signed-off-by: Tony Asleson <tasleson@redhat.com>