Specify size in KiB instead of bytes
LVM does some preliminary check that the given size is a multiple of 512 bytes
and exits with an error in case its not even though it would later just round it
to a multiple of extent size. See
https://bugzilla.redhat.com/show_bug.cgi?id=1221247 for more details.

To overcome this issue, let's just specify sizes in KiB using the 'K' suffix.