Fix intermittent fail in AllocationEnabled
AllocationEnabled was working most of the time but would fail
intermittently if the udev event came in a little late.  The
root issue is the Vg object sequence count does indeed get
incremented because of the meta data update.  Added an explicit
refresh to ensure this is done before we return.

Signed-off-by: Tony Asleson <tasleson@redhat.com>