pvmover.py: Update state after move
Previously when a move completed we did not update the state of the dbus
objects at all.  Thus with this change when the move is completed, but
before we return status to the user we update the state first.  To do this
we are placing a request into the work queue so that it executes in the
same context as other dbus requests as the mover executes in a separate
thread.

Signed-off-by: Tony Asleson <tasleson@redhat.com>