loader.py: Make sure object would be created the same
When we compare object state we also need to consider that the object path
and the object type that would be getting created.  This is because some
objects can change interfaces which causes an object instance change which
could cause the object path to be different as well.  If these do not match
we will then remove the old object from dbus and register a new one.

Signed-off-by: Tony Asleson <tasleson@redhat.com>