request.py: Return the stacktrace on code bug
Previously if the code hit an error and raised an exception the
service would catch it and log it while the client would just
time-out waiting for a response.  With this change we will grab
the stacktrace and return it as an error immediately.  This should
be more helpful for the client and for debugging the issue later.

Signed-off-by: Tony Asleson <tasleson@redhat.com>