objectmanager.py: Correct object id lookups
We were correctly identifying when we needed to correct an object
lvm_id or uuid.  However, we incorrectly removed the look-ups and
added the exact same old ones back which defeated the purpose of
identifying the situation to begin with.  Pass the new lvm_id and
uuid to the update method so that we can correctly add look-ups
for the change.

Signed-off-by: Tony Asleson <tasleson@redhat.com>