objectmanager.py: Ensure look-ups are valid
LVM has no identifier that is persistent and non-changing.  VG/LV names
and UUIDs can be changed which makes it difficult to uniquely identify
something in LVM over time.  The dbus service makes the assumption that
you can only change one or the other at a time and leverages this in
the objectmanager table lookups.  When we go to look for an object, we
pass it the uuid and lvm id (PV path, VG Name, LV full name).  We first
check to see if we can find it by lvm id.  If that succeeds we return
the object path, otherwise we do a uuid lookup.  If that fails the
object is not known.

Object look-ups are handled with two hash tables.  One by lvm id, the
other by uuid.  When a lvm id gets renamed or a uuid gets changed, we
are in a state where one look up is good and the other is stale.
This change verifies and updates the other look-up if:

a. The uuid and the lvm id are different
   (We sometimes only have one and look-up by it)
b. We fail to look-up using the other key

Signed-off-by: Tony Asleson <tasleson@redhat.com>